requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
mlflow==2.4
cloudpickle==2.2.1
numpy==1.23.5
psutil==5.9.5
scikit-learn==1.3.0
scipy==1.10.1
